% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{.finsys_dbh_arsendring}
\alias{.finsys_dbh_arsendring}
\title{Inneholder 책rsendring for berigning i Bl책tthefte}
\usage{
.finsys_dbh_arsendring()
}
\value{
a tibble budsjettar, arsendring_pst
}
\description{
Funksjon returnerer en vektor med 책rsendring
per budsjett책r
}
\keyword{internal}
