% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{.finsys_dbh_data}
\alias{.finsys_dbh_data}
\title{Finsieringsystemet data}
\usage{
.finsys_dbh_data(arstall)
}
\arguments{
\item{arstall}{a tibble}
}
\value{
a tibble
}
\description{
Filtrerer finsystabeller
til hva som gir uttelling, og harmoniserer
variabelnavn og -verdier til sammenslåing
}
\keyword{internal}
