% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{.finsys_dbh_doktorgarder}
\alias{.finsys_dbh_doktorgarder}
\title{DBH_API doktorgrader
Help function for doktorgrader kandidater}
\usage{
.finsys_dbh_doktorgarder(.data, finsys_dbh)
}
\arguments{
\item{.data}{a tibble}

\item{finsys_dbh}{a tibble}
}
\value{
a tibble
}
\description{
DBH_API doktorgrader
Help function for doktorgrader kandidater
}
\keyword{internal}
