% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{.finsys_dbh_ekonomi}
\alias{.finsys_dbh_ekonomi}
\title{DBH-API økonomidata}
\usage{
.finsys_dbh_ekonomi(.data, finsys_dbh)
}
\arguments{
\item{.data}{a tibble}

\item{finsys_dbh}{a tibble}
}
\value{
a tibble
}
\description{
Hjelp funksjon for økonomi data
}
\keyword{internal}
