% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{.finsys_dbh_indikatorsatser}
\alias{.finsys_dbh_indikatorsatser}
\title{Inneholder satser for resultatbaser uttelling og rammer data i Blåtthefte}
\usage{
.finsys_dbh_indikatorsatser()
}
\value{
a vector of institutions
}
\description{
Funksjon returnerer en vektor med satser og rammer
verdi for indikator og
budsjettår
}
\keyword{internal}
