% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{.finsys_dbh_institusjoner}
\alias{.finsys_dbh_institusjoner}
\title{Inneholder institusjoner i Blåtthefte}
\usage{
.finsys_dbh_institusjoner()
}
\value{
a vector of institutions
}
\description{
Funksjon returnerer en vektor med
institusjonskode
}
\keyword{internal}
