% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{.finsys_dbh_institutions}
\alias{.finsys_dbh_institutions}
\title{DBH_API institusjoner}
\usage{
.finsys_dbh_institutions(.data, finsys_dbh)
}
\arguments{
\item{.data}{a tibble}

\item{finsys_dbh}{a tibble}
}
\value{
a tibble
}
\description{
Hjelper funksjon med institusjoner fra DBH API
}
\keyword{internal}
