% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{.finsys_dbh_kandidater}
\alias{.finsys_dbh_kandidater}
\title{DBH-kandidater}
\usage{
.finsys_dbh_kandidater(.data, finsys_dbh)
}
\arguments{
\item{.data}{a tibble}

\item{finsys_dbh}{a tibble}
}
\value{
a tibble
}
\description{
Hjelp funksjon for kandidater
}
\keyword{internal}
