% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{.finsys_dbh_merge}
\alias{.finsys_dbh_merge}
\title{Finsieringsystemet indikator verdier}
\usage{
.finsys_dbh_merge(arstall)
}
\arguments{
\item{arstall}{a tibble}
}
\value{
a tibble
}
\description{
lager indikator verdier
}
\keyword{internal}
