% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{.finsys_dbh_query}
\alias{.finsys_dbh_query}
\title{DBH-tabeller med API-valg}
\usage{
.finsys_dbh_query(arstall, institusjoner = .finsys_dbh_institusjoner())
}
\arguments{
\item{arstall}{velg årstall}

\item{institusjoner}{institusjoner som delta i Blåttfete}
}
\value{
tibble
}
\description{
Spøring til DBH-API med en
liste over tabller i Blåtthefte
}
\keyword{internal}
