% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{.finsys_dbh_studiepoeng}
\alias{.finsys_dbh_studiepoeng}
\title{DBH_API studiepoeng}
\usage{
.finsys_dbh_studiepoeng(.data, finsys_dbh)
}
\arguments{
\item{.data}{a tibble}

\item{finsys_dbh}{a tibble}
}
\value{
a tibble
}
\description{
Hjelp funksjon for studiepoeng
}
\keyword{internal}
