% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{.finsys_dbh_tilskuddsgrad}
\alias{.finsys_dbh_tilskuddsgrad}
\title{Inneholder institusjoner med tilskuddsgrad i Blåtthefte}
\usage{
.finsys_dbh_tilskuddsgrad()
}
\value{
a vector of institutions
}
\description{
Funksjon returnerer en vektor med tilskuddsgrad
verdi for indikator og
institusjonskode
}
\keyword{internal}
