% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{.finsys_dbh_unntak}
\alias{.finsys_dbh_unntak}
\title{Inneholder unntakk institusjoner for indikator i Blåtthefte}
\usage{
.finsys_dbh_unntak()
}
\value{
a vector of institutions
}
\description{
Funksjon returnerer en vektor med unntak indikator og
institusjonskode
}
\keyword{internal}
