% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_basic_data.R
\name{finsys_budsjettendringer}
\alias{finsys_budsjettendringer}
\title{Lager budsjettendring tabell}
\usage{
finsys_budsjettendringer(df)
}
\arguments{
\item{df}{data set fra finsys_dbh_tidy}
}
\value{

}
\description{
budsjettendring tabeller fra api data
}
