% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{finsys_dbh_tidy}
\alias{finsys_dbh_tidy}
\title{Legger til endringstall fra året før}
\usage{
finsys_dbh_tidy(arstall)
}
\arguments{
\item{arstall}{a tibble}
}
\value{
a tibble
}
\description{
lager indikatorendring verdier
}
