% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_basic_data.R
\name{finsys_produksjonstabeller}
\alias{finsys_produksjonstabeller}
\title{Lager tabeller med produksjonsdata, tilsvarende dem i Blått hefte}
\usage{
finsys_produksjonstabeller(arstall)
}
\arguments{
\item{arstall}{velg årstall}
}
\value{
a tibble
}
\description{
funksjon for å lage tabeller med grunnlagsdata
}
