% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_basic_data.R
\name{lag_produksjonstabell}
\alias{lag_produksjonstabell}
\title{Lager tabeller med produksjonsdata, tilsvarende dem i Blått hefte}
\usage{
lag_produksjonstabell(df, filter_indikatorer, kolonne_vars, funs)
}
\arguments{
\item{df}{dataramme}

\item{filter_indikatorer}{indikatorene som skal inngå i tabellen}

\item{kolonne_vars}{variablene som skal spres på kolonner (indikator/kategori/osv.)}

\item{funs}{tilleggsfunksjon som skal anvendes på kolonnene (for å lage andelstall)}
}
\value{
a tibble
}
\description{
Lager tabeller med produksjonsdata, tilsvarende dem i Blått hefte
}
