% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{.finsys_dbh_prisjustering}
\alias{.finsys_dbh_prisjustering}
\title{Prisjustering}
\usage{
.finsys_dbh_prisjustering(.data)
}
\arguments{
\item{.data}{a tibble}
}
\value{
a tibble
}
\description{
Hjep funksjon for beregning av årsdeflator
}
\keyword{internal}
