% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{finsys_dbh_tidy}
\alias{finsys_dbh_tidy}
\title{Lager samletabell og beregner endringer og uttelling}
\usage{
finsys_dbh_tidy(arstall, .data)
}
\arguments{
\item{arstall}{velg årstall}

\item{.data}{a tibble}
}
\value{
a tibble
}
\description{
Bruker import data fra DBH-API og hjelper funksjoner
for å lage en samletabell med kategori, kandidatgruppe, faktor,
idikatorverdi, indikatorendring, budsjettendring som skal brukes
videre for å lage produksjonstableller
}
