% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{finsys_dbh_tidy}
\alias{finsys_dbh_tidy}
\title{Finsieringsystemet indikator og indikatorendring verdier}
\usage{
finsys_dbh_tidy(arstall)
}
\arguments{
\item{arstall}{a tibble}
}
\value{
a tibble
}
\description{
Legger til endringstall fra året før
}
\keyword{internal}
